feat: add GradBacDETR and BacDETRRecaller model variants
- add GradBacDETRConfig and BacDETRRecallerConfig with specialized defaults for gradient-adapter and recall-optimized segmentation
- add GradBacDETR and BacDETRRecaller model classes and export them from package
- fix dtype mismatch in IoU loss targets to prevent training errors

+++ bacdetr/bacdetr/models/lwdetr.py
@@ -401,7 +401,7 @@ class SetCriterion(nn.Module):
 
             pos_ind=[id for id in idx]
             pos_ind.append(target_classes_o)
-            cls_iou_func_targets[pos_ind] = pos_ious_func
+            cls_iou_func_targets[pos_ind] = pos_ious_func.to(cls_iou_func_targets.dtype)
             norm_cls_iou_func_targets = cls_iou_func_targets \
                 / (cls_iou_func_targets.view(cls_iou_func_targets.shape[0], -1, 1).amax(1, True) + 1e-8)
             loss_ce = position_supervised_loss(src_logits, norm_cls_iou_func_targets, num_boxes, alpha=self.focal_alpha, gamma=2) * src_logits.shape[1]
@@ -420,7 +420,7 @@ class SetCriterion(nn.Module):
 
             pos_ind=[id for id in idx]
             pos_ind.append(target_classes_o)
-            cls_iou_targets[pos_ind] = pos_ious
+            cls_iou_targets[pos_ind] = pos_ious.to(cls_iou_targets.dtype)
             loss_ce = sigmoid_varifocal_loss(src_logits, cls_iou_targets, num_boxes, alpha=self.focal_alpha, gamma=2) * src_logits.shape[1]
         else:
             target_classes = torch.full(src_logits.shape[:2], self.num_classes,
